refactor(sitesservice): use ts
